+++ src/riverqueue/driver/__init__.py
@@ -1,7 +1,7 @@
 # Reexport for more ergonomic use in calling code.
 from .driver_protocol import (
     ExecutorProtocol as ExecutorProtocol,
-    GetParams as GetParams,
+    JobGetByKindAndUniquePropertiesParam as JobGetByKindAndUniquePropertiesParam,
     JobInsertParams as JobInsertParams,
     DriverProtocol as DriverProtocol,
 )